WatchlistCard refactored
