+++ src/components/Watchlists/NewWatchlistCard.js
@@ -3,7 +3,8 @@ import cx from 'classnames'
 import LoginPopup from '../banners/feature/PopupBanner'
 import { useUserWatchlists } from '../../ducks/Watchlists/gql/hooks'
 import NewWatchlist from '../../ducks/Watchlists/Actions/New'
-import styles from './../../ducks/Watchlists/Cards/WatchlistCard.module.scss'
+import styles from './../../ducks/Watchlists/Cards/NewCard.module.scss'
+import cardStyles from './../../ducks/Watchlists/Cards/Card.module.scss'
 
 export const SvgNew = ({ className }) => (
   <svg
@@ -44,7 +45,7 @@ export const SvgNew = ({ className }) => (
 
 const Trigger = ({ title, ...rest }) => {
   return (
-    <div className={cx(styles.wrapper, styles.create)} {...rest}>
+    <div className={cx(cardStyles.wrapper, styles.create)} {...rest}>
       <SvgNew />
       <div className={styles.createLink}>Create your watchlist</div>
     </div>
