+++ src/ducks/Watchlists/Cards/Card.js
@@ -2,14 +2,14 @@ import React, { useMemo } from 'react'
 import cx from 'classnames'
 import gql from 'graphql-tag'
 import { useQuery } from '@apollo/react-hooks'
-import emptyChartSvg from './Cards/emptyChart.svg'
-import { getSEOLinkFromIdAndTitle } from '../../utils/url'
-import { calcPercentageChange } from '../../utils/utils'
-import { millify } from '../../utils/formatting'
-import MiniChart from '../../components/MiniChart'
-import PercentChanges from '../../components/PercentChanges'
-import NewLabel from '../../components/NewLabel/NewLabel'
-import { VisibilityIndicator } from '../../components/VisibilityIndicator'
+import emptyChartSvg from './emptyChart.svg'
+import { getSEOLinkFromIdAndTitle } from '../../../utils/url'
+import { calcPercentageChange } from '../../../utils/utils'
+import { millify } from '../../../utils/formatting'
+import MiniChart from '../../../components/MiniChart'
+import PercentChanges from '../../../components/PercentChanges'
+import NewLabel from '../../../components/NewLabel/NewLabel'
+import { VisibilityIndicator } from '../../../components/VisibilityIndicator'
 import styles from './Card.module.scss'
 
 export const WATCHLIST_MARKETCAP_HISTORY_QUERY = gql`
