+++ src/ducks/Watchlists/Cards/Card.module.scss
