+++ src/ducks/Watchlists/Cards/MyWatchlist.js
@@ -1,8 +1,7 @@
 import React from 'react'
 import cx from 'classnames'
 import Button from '@santiment-network/ui/Button'
-import WatchlistCard from './WatchlistCard'
-import { getWatchlistLink } from '../utils'
+import { WatchlistCards } from './Card'
 import { DesktopOnly, MobileOnly } from './../../../components/Responsive'
 import EmptySection from '../../../components/EmptySection/EmptySection'
 import Skeleton from '../../../components/Skeleton/Skeleton'
@@ -82,16 +81,8 @@ const MyWatchlist = ({ className, showHeader = true, classes = {} }) => {
       )}
       {isLoggedIn && (
         <div className={stylesGrid.wrapper}>
-          {watchlists.map(watchlist => (
-            <WatchlistCard
-              key={watchlist.id}
-              name={watchlist.name}
-              watchlist={watchlist}
-              to={getWatchlistLink(watchlist)}
-              isPublic={watchlist.isPublic}
-              {...watchlist}
-            />
-          ))}
+          <WatchlistCards watchlists={watchlists} path='/watchlist/projects/' />
+
           {watchlists.length > 0 && <NewWatchlistCard />}
         </div>
       )}
