--- src/ducks/Watchlists/Cards/WatchlistCard.module.scss
@@ -1,161 +1,0 @@
-@import '~@santiment-network/ui/mixins';
-
-.wrapper {
-  @include text-normal;
-
-  border: 1px solid var(--porcelain);
-  border-radius: $border-radius;
-  display: block;
-  padding: 16px 20px;
-  height: 120px;
-  position: relative;
-  overflow: hidden;
-  color: var(--rhino);
-  cursor: pointer;
-
-  @include responsive('laptop', 'desktop') {
-    &:hover {
-      border-color: var(--mystic);
-
-      .name {
-        color: var(--jungle-green-hover);
-      }
-    }
-  }
-
-  @include responsive('phone', 'phone-xs') {
-    padding: 14px 16px;
-    margin-top: 8px;
-  }
-
-  &__simple {
-    padding: 12px 20px 12px 16px;
-    margin-top: 8px;
-    height: auto;
-  }
-}
-
-.name {
-  line-height: 20px;
-  width: 100%;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  color: var(--rhino);
-
-  &::first-letter {
-    text-transform: uppercase;
-  }
-
-  @include responsive('phone', 'phone-xs') {
-    @include text('body-2');
-  }
-}
-
-.flexRow {
-  display: flex;
-  align-items: center;
-
-  @include responsive('phone', 'phone-xs') {
-    @include text('body-2');
-  }
-}
-
-.content {
-  justify-content: space-between;
-}
-
-.volumeLabel {
-  color: var(--casper);
-}
-
-.change {
-  margin-left: 0;
-  font-size: 14px;
-
-  @include responsive('phone', 'phone-xs') {
-    @include text('body-2');
-  }
-
-  &__simple {
-    margin: 0 0 0 12px;
-    white-space: nowrap;
-  }
-}
-
-.marketcap {
-  margin: 12px 0 8px;
-
-  @include text('h4');
-
-  color: var(--rhino);
-
-  @include responsive('phone', 'phone-xs') {
-    margin: 10px 0 4px;
-  }
-}
-
-.chart {
-  max-width: 100px;
-
-  @include responsive('laptop', 'desktop', 'tablet') {
-    width: 65px;
-  }
-}
-
-.new {
-  margin-right: 8px;
-}
-
-.createLink {
-  margin-left: 17px;
-}
-
-.create {
-  border: 1px solid var(--porcelain);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: var(--rhino);
-
-  @include text('body-2');
-
-  &:hover {
-    border: 1px solid var(--jungle-green-light-3);
-
-    .createLink {
-      color: var(--jungle-green);
-    }
-  }
-
-  &__disabled {
-    color: var(--waterloo);
-
-    &:hover {
-      border: 1px solid var(--porcelain);
-
-      .createLink {
-        color: var(--waterloo);
-      }
-    }
-  }
-}
-
-.proLabel {
-  position: absolute;
-  top: 8px;
-  right: 8px;
-}
-
-.loader {
-  font-size: 10px;
-}
-
-.link {
-  color: var(--texas-rose);
-  display: block;
-
-  &:hover {
-    color: var(--texas-rose-hover);
-  }
-}
