+++ src/pages/Watchlists/index.js
@@ -5,8 +5,8 @@ import { useUser } from '../../stores/user'
 import { DesktopOnly, MobileOnly } from '../../components/Responsive'
 import StoriesList from '../../components/Stories/StoriesList'
 import RecentlyWatched from '../../components/RecentlyWatched/RecentlyWatched'
-import { WatchlistCards } from '../../ducks/Watchlists/Card'
-import FeaturedWatchlistCards from '../../ducks/Watchlists/Featured'
+import { WatchlistCards } from '../../ducks/Watchlists/Cards/Card'
+import FeaturedWatchlistCards from '../../ducks/Watchlists/Cards/Featured'
 import { WatchlistEmptySection } from '../../ducks/Watchlists/Cards/MyWatchlist'
 import {
   useUserWatchlists,
